+++ include/ambition/7kaaInterface/draw.hh
@@ -373,6 +373,12 @@ bool printSaveFilename(
   const int top
 );
 
+bool villagesReportTownPopulations(
+  const int top,
+  const int villagers,
+  const int peasants
+);
+
 bool villagesReportUpperTotals(
   const int totalVillagers,
   const int totalPeasants