+++ src/ambition/Ambition_vga.cc
@@ -1822,6 +1822,41 @@ void printMilitaryReportTotalCost(
   );
 }
 
+void printVillagesReportTownPopulations(
+  const int top,
+  const int villagers,
+  const int peasants
+) {
+  using namespace UserInterface::Report::Villages::Towns;
+
+  const auto rowArea = UserInterface::Rectangle::fromPoint(
+    {
+      .left = AREA.start.left,
+      .top = top,
+    },
+    {
+      .width = AREA.width(),
+      .height = 21,
+    }
+  );
+
+  UserInterface::printText(
+    font_san,
+    format("%'d", villagers),
+    rowArea.intersection(VILLAGERS_COLUMN),
+    UserInterface::Clear::None,
+    UserInterface::HorizontalAlignment::Right
+  );
+
+  UserInterface::printText(
+    font_san,
+    format("%'d", peasants),
+    rowArea.intersection(PEASANTS_COLUMN),
+    UserInterface::Clear::None,
+    UserInterface::HorizontalAlignment::Right
+  );
+}
+
 void printVillagesReportUpperTotals(
   const int totalVillagers,
   const int totalPeasants,