+++ src/ambition/version_details.cc
@@ -42,6 +42,7 @@ std::vector<VersionUpdates> versionDetails = {
         .items = {
           _("Increased bribe choices to include $2500 and $3500."),
           _("Made the build guidelines on the Standard tileset slightly brighter, to make them more visible."),
+          _("Lined up the villager and peasant numbers on the Villages Report."),
         },
       },
     },